+++ ChangeLog.md
@@ -12,6 +12,9 @@
 * **[new]** Adds `RandomExponential`, `RandomGaussian`, and `RandomUniform` to
   `lib/math` for generating random `Float` values.
 
+* **[new]** Adds `Realtime.sleepSecondsPrecise` function to allow sleeps that
+  are more precise than the kernel's latency.
+
 ## 0.18.0.0  -- 2021-08-01
 
 ### Compiler CLI